+++ lib/ui/home/HomeCampusHighlightsWidget.dart
@@ -7,6 +7,7 @@ import 'package:illinois/service/Config.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/home/HomeWidgets.dart';
 import 'package:illinois/ui/widgets/LinkButton.dart';
+import 'package:illinois/ui/widgets/SemanticsWidgets.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
@@ -140,6 +141,7 @@ class _HomeCampusHighlightsWidgetState extends State<HomeCampusHighlightsWidget>
           key: _pageViewKey,
           controller: _pageController,
           estimatedPageSize: _pageHeight,
+          allowImplicitScrolling: true,
           children: pages,
         ),
       );
@@ -153,6 +155,7 @@ class _HomeCampusHighlightsWidgetState extends State<HomeCampusHighlightsWidget>
 
     return Column(children: <Widget>[
       contentWidget,
+      AccessibleViewPagerNavigationButtons(controller: _pageController, pagesCount: visibleCount,),
       LinkButton(
             title: Localization().getStringEx('widget.home.campus_guide_highlights.button.all.title', 'View All'),
             hint: Localization().getStringEx('widget.home.campus_guide_highlights.button.all.hint', 'Tap to view all highlights'),