+++ lib/ui/home/HomeGroupsWidget.dart
@@ -7,6 +7,8 @@ import 'package:illinois/ui/groups/GroupsHomePanel.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/home/HomeWidgets.dart';
 import 'package:illinois/ui/widgets/LinkButton.dart';
+import 'package:illinois/ui/widgets/SemanticsWidgets.dart';
+import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/group.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
 import 'package:rokwire_plugin/service/auth2.dart';
@@ -15,6 +17,7 @@ import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:illinois/ui/groups/GroupWidgets.dart';
+import 'package:rokwire_plugin/service/styles.dart';
 
 
 class HomeMyGroupsWidget extends StatefulWidget {
@@ -123,8 +126,10 @@ class _HomeMyGroupsState extends State<HomeMyGroupsWidget> implements Notificati
       for (Group? group in _groups!) {
         if (group != null) {
           pages.add(Padding(padding: EdgeInsets.only(right: _pageSpacing), child:
-            GroupCard(group: group, displayType: GroupCardDisplayType.homeGroups, margin: EdgeInsets.zero,),
-          ));
+            Semantics(
+              // excludeSemantics: !(_pageController?.page == _groups?.indexOf(group)),
+             child: GroupCard(group: group, displayType: GroupCardDisplayType.homeGroups, margin: EdgeInsets.zero,),
+          )));
         }
       }
     }
@@ -143,8 +148,10 @@ class _HomeMyGroupsState extends State<HomeMyGroupsWidget> implements Notificati
           key: _pageViewKey,
           controller: _pageController,
           children: pages,
+          allowImplicitScrolling : true,
         )
       ),
+      AccessibleViewPagerNavigationButtons(controller: _pageController, pagesCount: pages.length,),
       LinkButton(
         title: Localization().getStringEx('widget.home.my_groups.button.all.title', 'View All'),
         hint: Localization().getStringEx('widget.home.my_groups.button.all.hint', 'Tap to view all groups'),
@@ -184,7 +191,6 @@ class _HomeMyGroupsState extends State<HomeMyGroupsWidget> implements Notificati
     }
     return HomeMessageCard(message: message,);
   }
-
 
   @override
   void onNotification(String name, param) {