+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add link to Wellness Resources Home when favorites widget is empty [#2098](https://github.com/rokwire/illinois-app/issues/2098).
 ### Changed
 - Make Wellness widget components stand alone [#2107](https://github.com/rokwire/illinois-app/issues/2107).
-
+### Fixed
+- Fix Horizontal scrolling is not accessible [#2093](https://github.com/rokwire/illinois-app/issues/2093).
 ## [4.1.3] - 2022-07-26
 ### Changed
 - Rename "Reminders" tab to "Weekly" in Wellness ToDo list [#1967](https://github.com/rokwire/illinois-app/issues/1967).